FIX: Change return type of getGeneLength to size_t; fix incorrect variable in Builder::setParameterLengths(); update error messages for parameter validation in Builder methods

+++ include/gaconfig.hpp
@@ -15,7 +15,7 @@ class GAConfig {
 
         // GETTERS
         
-        double getGeneLength();
+        size_t getGeneLength();
         size_t getNumParameters();
         std::vector<size_t> getParameterLengths();
         std::vector<std::pair<double, double>> getParameterRanges();
